requirements.txt
beautifulsoup4 @ file:///C:/b/abs_0agyz1wsr4/croot/beautifulsoup4-split_1681493048687/work
chroma-hnswlib==0.7.3
chromadb==0.4.22
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langchain-openai==0.0.5
langsmith==0.0.86
openai==1.10.0
python-dotenv==1.0.1
